Add Senor and Platform classes and Range & Bearing Model
RangeBearingGaussianToCartesian: Non-linear model which transforms between 2D Cartesian and 2D Polar coordinates.

Sensor & SimpleRadar: The Sensor class is the base class for all sensors, while SimpleRadar provides an implementation of a simple Radar sensor that utilises the RangeBearingGaussianToCartesian measurement model, in order to generate measurements of targets.

Platform & SensorPlatform: The Platform class is the base class for all platforms and allows for the use of a TransitionModel, which, when specified, enables for the simulation of platform motion. The SensorPlatform class, extends the base Platform class and allows for a variable number of different sensors to be mounted on the platform, at different offsets, whose positions are automatically updated as the platform moves in space.