Fix double "https://" in README link to utm Lib
This commit aims to fix an error in the README file. In the section about "Dependencies" the protocol in the link to the utm library was defined twice. It is a minor bug, but it may lead to inconvenience for users trying to explore if the stonesoup project is a good option for them.