Added Nth derivative decay model
This is a model that allows an arbitrary derivative to decay to 0
exponentially akin to singer and OU. It also makes singer and OU a
branch off the Nth derivative decay model.

Notes:
Due to small deviations from the original values, the tests for the
singer model were adapted to use AllClose instead of AllEqual the
tolerance is 1e-6, so the deviations are minuscule.
It is possible to add a similar generalisation for the approximate
singer model, but I wasn't sure if this was necessary. If anyone would
like it, let me know.