Add AESA radar sensor and refactor radar sensors
Added an AESA (Active electronically scanned array) radar sensor along
with functions for the movement of the radar beam and the beam shape.
The radar sensor calculates the probability of detection of a state and
generates a measurement based on that probability.

This also refactors radar sensors, to allow more complex radar sensors
to be made, by customising radar beam shape and patterns.

Closes #156 (squash merge)