Change model functions to accept State not StateVector
This changes the behaviour of the `model`.`function()` function to take
a state, rather than a state vector. `function()` still returns a state
vector and `pdf()` still returns the value of the pdf. This allows
`function()` to use other parameters in the state, not just those
contained in the state vector.

Closes #155 (squash merge)