Add StateVectors type and modify array handling
This includes a number of changes to how Matrix type is implemented
which will allow greater customisation if required. In particular, a new
StateVectors type has been created which handles the case where
calculations of averages (and in turn covariance) were incorrect when
using Angle types.

Angle types now include a average method, which returns a circular mean.
This removes an issue when calculating means near the point of angle
wrapping.

These changes were primarily done for bug identified when calculating
mean of particles, and such some minor changes to PartcileState and
Particle type have been made.

Some functions have also been updated, as these are a set of state
vectors, so can take advantage of the new type. This includes updating
the Mixture Tracker

The array changes have changed minimum requirement for NumPy to 1.17, or
1.16 with NUMPY_EXPERIMENTAL_ARRAY_FUNCTION environment variable set to
1.