Fix caching issue with states/mutable-sequences with predictor
When passing a `StateMutableSequence` (e.g. a track) to a predictor
using caching, the cache can incorrectly return results if same track
and time stamp are passed in multiple times. This can be the case with
multi-sensor example, where an update has occurred from one sensor, then
a detection from another sensor is processed at the same time stamp.

This fixes the issue by using a custom lru_cache decorator that extracts
the current state if input is mutable sequence.