Add smooth transition model maker
This branch adds the 'get_smooth_transition_models' method.
This is to provide more realistic transition models/times for the 'MultiTransitionMovingPlatform' class when simulating motion between
given 2D cartesian coordinates.

Note:
The mathematics in the model maker are not using meaningful variable
names for the sake of not cluttering up the code. Instead I've added
in-line comments where needed. Alterations to this are welcome.