Added optional second mapping for all measures. If not None, then this will apply a second different mapping to state2 than to state1. If it is None, then it takes the same value as mapping, which is functionally equivalent to what we were doing before (i.e. one map for both states).
It will return an error if mapping2 is set but mapping is None. This is in theory something a user might want to do, so the error contains instructions for how to achieve this (i.e. set mapping to include all dimensions)

Added optional second position and velocity mapping to the SIAPs metric. It will default to being the same as the first mapping if not set. It will return an error message if position_mapping2 or velocity_mapping2 is set but the first mappings are None.