Correct bug when dealing with duplicate tags
When multiple states have the same tag, we want to keep that tag with the state that has the highest weight. The rest of the states will get new tags. Previously it accidently kept the tag with the state with the lowest weight. To fix this, we just have to sort the components from highest to lowest weight instead of the default lowest to highest.