Add method to copy prior state properties of a platform across to a new state when moving
The motivation behind this change is to allow the use of GroundTruthState types in platforms.
Previously, a platform's (movement controller's) `move` method hard-coded a standard `State` append.
This lead to loss of metadata when initiating a platform with a GroundTruthState, since the metadata property was not copied across to the new state.