Add Nonparametric PDA and gate out non valid measurements
This adds the ability to use nonparametric PDA for clutter spatial
density, where it is estimated based on assumption that measurements
within a defined validation region represent the clutter level.

This change also gates out non valid measurements by default, similar to
the behaviour in the DistanceHypothesier, and per description in
original PDA paper.