Add caching of mean and covar on ParticleState
This has significant performance improvements, for the trade off of
limiting editing to state vector and weight of the ParticleState.
However, they can still be replaced with another array, clearing the
cache.

Minor issue is use of standard library `functools.cached_property`
is only available from Python 3.8, so to maintain compatibility with
Python 3.7, `cached_property` code snippet is also included.