Change inequality due to is_clutter_detectable
Certain sensors have a FOV where the clutter measurement space can be
calculated as to if the clutter is within the FOV or not, as such,
cluter measurements can be generated outside the FOV and thus removed
from the set of clutter measurements. This means a single TrueDetection
combined with a set of clutter measurements where all are outside the
FOV returns a single TrueDetection.