Allow generation of predicted measurement without measurement noise
This standardises the Updater.predict_measurement method to include an
optional measurement_noise boolean argument to whether noise should be
included in the predicted measurement. Defaults to True.

Note this does change the default behaviour of particle filter, as that
did not included noise by default before. As this isn't used in the
update step, it'll only have an impact on data association stage.

Resolves #961